MAINTAINERS: update maintain status of FIRMWARE LOADER
Recently firmware loader code has been cleaned up much and
cache/uncache firmware mechanism is introduced, so later
drivers should apply the mechanism to solve the problem
of firmware lost during suspend/resume cycle.

For the sake of helping the related problems, I'd like to
volunteer to maintain the code and help on drivers' related
problem, so update the item.

Signed-off-by: Ming Lei <ming.lei@canonical.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>